task3/requirements.txt
celery
python-dotenv
requests
bs4
redis